refactor: 경매 상태 관리 enum 클래스 추가(#33)
