+++ src/main/java/com/skyhorsemanpower/auction/common/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.skyhorsemanpower.auction.common;
 
+import com.skyhorsemanpower.auction.status.ResponseStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
